Add files via upload
The HTML file 'index.html' is the backbone of a Weather App project. It establishes the structure and content of the web page, integrating essential elements for user interaction. Through HTML5 markup, it defines the layout, including search functionality, weather details display areas, and placeholders for dynamic content. It sets the stage for a responsive and visually appealing interface, paving the way for a seamless user experience in exploring and retrieving weather information. This HTML file serves as the foundation for combining design and functionality, forming the core structure of the Weather App.